chore(project): 파일 확장자 타입스크립트로 변경
