+++ package.json
@@ -20,15 +20,19 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/node": "^18.11.18",
-    "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
-    "@types/styled-components": "^5.1.34"
+    "@craco/craco": "^7.1.0",
+    "@craco/types": "^7.1.0",
+    "@types/react": "^19.0.5",
+    "@types/styled-components": "^5.1.34",
+    "tsconfig-paths-webpack-plugin": "^4.2.0",
+    "webpack-bundle-analyzer": "^4.10.1"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "craco start",
+    "build": "craco build",
+    "test": "craco test",
+    "eject": "craco eject",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
   },
@@ -50,10 +54,5 @@
       "last 1 safari version"
     ]
   },
-  "homepage": "https://kiimdohyun.github.io/KiimDoHyun-portfolio_window",
-  "devDependencies": {
-    "@types/react": "^19.0.5",
-    "@types/styled-components": "^5.1.34",
-    "webpack-bundle-analyzer": "^4.10.1"
-  }
+  "homepage": "https://kiimdohyun.github.io/KiimDoHyun-portfolio_window"
 }
