+++ src/Component/TaskBar/HiddenIcon.tsx
