+++ src/Component/TaskBar/InfoBar.tsx
