+++ src/Component/TaskBar/SkillIcon.tsx
