+++ src/Component/TaskBar/StatusBar/CenterAreaBox.tsx
