+++ src/Component/TaskBar/StatusBar/LeftAreaBox.tsx
