+++ src/store/global.ts
