+++ src/store/taskbar.ts
